% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_notes.R
\docType{data}
\name{Pak_Indicators_Data}
\alias{Pak_Indicators_Data}
\title{PAKISTAN Indicators data}
\format{
## `Pak_Indicators_Data`
A data frame (long format) with 270128 rows and 15 columns:
\describe{
  \item{district}{District name}
  \item{province}{Province Name}
  \item{indicator} {Particular indicator}
  \item{year} {Year of the survey}
  \item{source} {source of the survey}
  \item{value}{Indicator Value}
  \item{definition} {Definition of the indicator}
  \item{unit} {Units of the indicator}
  ...
}
}
\source{
Author's calculations based on the respective surveys
}
\usage{
Pak_Indicators_Data
}
\description{
Pakistan Indicators finalized data, excluding AJK and GB
}
\keyword{datasets}
