% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_notes.R
\docType{data}
\name{Pak_Shapfiles}
\alias{Pak_Shapfiles}
\title{PAKISTAN SHAPEFILE 2022 data}
\format{
## `Pak_Shapfiles`
A data frame with 136 rows and 5 columns:
\describe{
  \item{district}{District name}
  \item{province}{Province Name}
  \item{geometry}{Geometries}
  ...
}
}
\source{
<https://data.humdata.org/dataset/cod-ab-pak?>
}
\usage{
Pak_Shapfiles
}
\description{
Pakistan Shapefiles excluding AJK and GB
}
\keyword{datasets}
